.ado/pipelines/python/CosmosDB/requirements.txt
azure-cosmos
azure-identity
azure-keyvault-secrets
